backend/PersonalReviewSystem/requirements.txt
django
djangorestframework
psycopg2
pygments
coreapi
django-auth-ldap
django-python3-ldap
markdown